requirements.txt
git+https://github.com/Rapptz/discord.py
asyncpg
python-decouple
python-dateutil
parsedatetime